fix casting of score

+++ src/pheval_phen2gene/post_process/post_process_results_format.py
@@ -26,7 +26,7 @@ def extract_gene_results(
             pl.col("Gene")
             .map_elements(gene_identifier_updator.find_identifier, return_dtype=pl.String)
             .alias("gene_identifier"),
-            pl.col("Score").alias("score"),
+            pl.col("Score").alias("score").cast(pl.Float64),
         ]
     )
